Password confirmation on user deletetion inertia

+++ src/Http/Controllers/Inertia/CurrentUserController.php
@@ -6,6 +6,8 @@ use Illuminate\Contracts\Auth\StatefulGuard;
 use Illuminate\Http\Request;
 use Illuminate\Routing\Controller;
 use Laravel\Jetstream\Contracts\DeletesUsers;
+use Illuminate\Support\Facades\Hash;
+use Illuminate\Validation\ValidationException;
 
 class CurrentUserController extends Controller
 {
@@ -18,6 +20,12 @@ class CurrentUserController extends Controller
      */
     public function destroy(Request $request, StatefulGuard $auth)
     {
+        if (! Hash::check($request->password, $request->user()->password)) {
+            throw ValidationException::withMessages([
+                'password' => [__('This password does not match our records.')],
+            ])->errorBag('deleteUser');
+        }
+
         app(DeletesUsers::class)->delete($request->user()->fresh());
 
         $auth->logout();
